Extract extension test util
