+++ .changelog/1768.internal.md
@@ -1,0 +1,1 @@
+Refactor extension tests
