+++ playwright/tests/extension.spec.ts
@@ -1,45 +1,10 @@
-import { test as base, expect, BrowserContext, chromium } from '@playwright/test'
-import path from 'path'
+import { test } from '../utils/extensionTestExtend'
+import { expect } from '@playwright/test'
 import { warnSlowApi } from '../utils/warnSlowApi'
 import { mockApi } from '../utils/mockApi'
 import { expectNoErrorsInConsole } from '../utils/expectNoErrorsInConsole'
 import { fillPrivateKeyWithoutPassword } from '../utils/fillPrivateKey'
 import { privateKey, privateKeyAddress } from '../../src/utils/__fixtures__/test-inputs'
-
-// Test dev build by default, but also allow testing production
-const extensionPath = path.join(__dirname, '..', process.env.EXTENSION_PATH ?? '../build-dev/')
-
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const extensionManifest = require(path.join(extensionPath, '/manifest.json'))
-const popupFile = extensionManifest.browser_action.default_popup
-
-// From https://playwright.dev/docs/chrome-extensions
-export const test = base.extend<{
-  context: BrowserContext
-  extensionId: string
-}>({
-  // eslint-disable-next-line no-empty-pattern
-  context: async ({}, use) => {
-    const context = await chromium.launchPersistentContext('', {
-      headless: false,
-      args: [`--disable-extensions-except=${extensionPath}`, `--load-extension=${extensionPath}`],
-    })
-    await use(context)
-    await context.close()
-  },
-  extensionId: async ({ context }, use) => {
-    // for manifest v2:
-    let [background] = context.backgroundPages()
-    if (!background) background = await context.waitForEvent('backgroundpage')
-
-    // for manifest v3:
-    // let [background] = context.serviceWorkers()
-    // if (!background) background = await context.waitForEvent('serviceworker')
-
-    const extensionId = background.url().split('/')[2]
-    await use(extensionId)
-  },
-})
 
 test.beforeEach(async ({ context }) => {
   await warnSlowApi(context)
@@ -47,14 +12,14 @@ test.beforeEach(async ({ context }) => {
 })
 
 test.describe('The extension popup should load', () => {
-  test('should successfully load javascript chunks', async ({ page, extensionId }) => {
-    await page.goto(`chrome-extension://${extensionId}/${popupFile}`)
+  test('should successfully load javascript chunks', async ({ page, extensionPopupURL }) => {
+    await page.goto(`${extensionPopupURL}/`)
     await expect(page.getByRole('link', { name: /Open wallet/i })).toBeVisible()
     await expect(page.getByRole('link', { name: /Create wallet/i })).toBeVisible()
   })
 
-  test('get state from background page through webext-redux', async ({ page, extensionId }) => {
-    await page.goto(`chrome-extension://${extensionId}/${popupFile}`)
+  test('get state from background page through webext-redux', async ({ page, extensionPopupURL }) => {
+    await page.goto(`${extensionPopupURL}/`)
     await page.getByRole('button', { name: /Dark mode/i }).click()
     await page.getByRole('button', { name: /Light mode/i }).click()
 
@@ -62,8 +27,8 @@ test.describe('The extension popup should load', () => {
     await expect(page.getByTestId('mnemonic-grid').locator('> *')).toHaveCount(24)
   })
 
-  test('ask for USB permissions in ledger popup', async ({ page, context, extensionId }) => {
-    await page.goto(`chrome-extension://${extensionId}/${popupFile}#/open-wallet`)
+  test('ask for USB permissions in ledger popup', async ({ page, context, extensionPopupURL }) => {
+    await page.goto(`${extensionPopupURL}/open-wallet`)
     const popupPromise = context.waitForEvent('page')
     await page.getByRole('button', { name: /Grant access to your Ledger/i }).click()
     const popup = await popupPromise
@@ -74,7 +39,7 @@ test.describe('The extension popup should load', () => {
     await expect(popup.getByText('error').or(popup.getByText('fail'))).toBeHidden()
   })
 
-  test('should allow embedded Transak widget', async ({ page, extensionId }) => {
+  test('should allow embedded Transak widget', async ({ page, extensionPopupURL }) => {
     await expectNoErrorsInConsole(page, {
       ignoreError: msg => {
         // Odd errors inside Transak
@@ -82,7 +47,7 @@ test.describe('The extension popup should load', () => {
         if (msg.text().includes('`sessionKey` is a required property')) return true
       },
     })
-    await page.goto(`chrome-extension://${extensionId}/${popupFile}#/open-wallet/private-key`)
+    await page.goto(`${extensionPopupURL}/open-wallet/private-key`)
     await fillPrivateKeyWithoutPassword(page, {
       privateKey: privateKey,
       privateKeyAddress: privateKeyAddress,
